requirements.txt
Flask==3.0.2
tensorflow==2.15.0
tf-keras==2.15.0
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
requests==2.31.0
sentence-transformers==2.6.1
scikit-learn==1.5.0
dotenv